backend/requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
qdrant-client
pyautogen
sqlalchemy
pydantic
psycopg2
dotenv
langchain_openai
passlib[bcrypt]
bcrypt==4.0.1
python-jose
python-multipart
cloudinary
python-dotenv
pydantic[email]
"unstructured[all-docs]"
